fixing remote build issue
